requirements.txt
nuscenes-devkit==1.1.11
numpy==1.26.4